ARM: layerscape: ppa: Fix starting PPA
With dba1c26f70 we replaced request_sdram_region() for the PPA with
reserve_sdram_region(). The effect is that the region is marked as
reserved and mapped non executable. While this is desired for EL2, it
also has the effect that we can't start the PPA anymore from EL3.

Map the region cached/executable to start the PPA, then map it
uncached/non executable once we are in EL2.

Fixes: dba1c26f70 ("arm: layerscape: ppa: reserve SDRAM region for PPA")
Link: https://lore.barebox.org/20231120144453.1075740-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>