ARM: stm32mp: don't re-enable DBGCFGR clock
TF-A will already enable the DBGMCU peripheral for us, so let's
skip setting the bit. This is useful, because while the peripheral's
base address stays the same between MP15 and MP13, the clock's offset
within the RCC peripheral doesn't, so by skipping the clock enablement,
we avoid having to differentiate between MP13 and MP15 that early.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231122181116.591131-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>