ARM: stm32mp: remove STM32MP13 .stm32 image
Legacy STM32 images were deprecated for STM32MP15 in TF-A and are no
longer supported in recent versions. For STM32MP13, FIP was the only
format supported since the beginning, so drop the unused entry point.

barebox chainloading for development can be done by booting the
barebox-stm32mp-generic-bl33.img with the barebox device tree passed
in externally.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231127064947.2207726-14-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>