ARM: i.MX7: Import GoMe device trees from kernel
The kernel device tree for the VAR-SOM-MX7 System-on-Module and the
Gossen Metrawatt Profitest MF (e143_01) have been submitted upstream:
http://lore.kernel.org/r/20231127-b4-imx7-var-som-gome-v1-0-f26f88f2d0bc@pengutronix.de

Until the fixes are in the upstream kernel and synced to our dts/
folder, add them here so we can already use them. This also includes the
dt-schema validation fixes to the upstream i.MX7 device trees mentioned
in the upstream kernel series.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20231127214818.3751722-2-rhi@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>