Documentation: user: barebox: make barebox chainloading clearer
barebox images are designed to be chainloadable and can be loaded as if
they were a kernel by other bootloaders. The documentation lists
examples for two ways to achieve this with U-Boot preinstalled on the
board. They are not as clear as they could be as the resulting uImage
is called image in the mkimage line, but barebox.bin later on. This is
unfortunate as the build results in a barebox.bin already, which
corresponds to barebox proper in PBL setups. Rename it to barebox.uImage
instead and while at it note that booti isn't applicable to arm32.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231129115255.1538958-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>