FIT: support kernel images with type = "kernel_noload"
U-Boot interprets "kernel_noload" to mean that the load and entry
addresses shall be ignored[1] and that the kernel image should be executed
in-place, unless compressed[2]. The entry and load addresses are still
mandatory and need to be initialized to some dummy value according to
spec[3].

barebox, which is unaware of any special semantics for the kernel_noload
type, would thus try to place a kernel_noload image at the dummy load
address specified and fail if that's not possible. Fix this by treating
type = "kernel_noload" as if load and entry properties were omitted,
in which case barebox falls back to find a suitable memory region at
runtime.

This change is motivated by the Linux kernel series adding FIT as
additional Kbuild target for ARM64[4]. With the change here, it's possible
to consume these FIT images in barebox as well.

[1]: U-Boot commit b9b50e89d317 ("image: Implement IH_TYPE_KERNEL_NOLOAD")
[2]: https://patchwork.ozlabs.org/project/uboot/list/?series=382849&state=*
[3]: https://github.com/open-source-firmware/flat-image-tree/releases/download/v0.8/fit-specification-v0.8.pdf
[4]: https://lore.kernel.org/linux-arm-kernel/20231129172200.430674-1-sjg@chromium.org/T/#meb5bda548de8d8d403c67ee90f639923c8a182fa

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231129203106.2417486-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>