kbuild: don't warn about reg formatting
Overlays, especially new style ones, are prone to trigger the reg
formatting warnings if fragments add node with reg properties without
also patching the parent:

  &phy0 {
	reg = <7>;
  };

In order to add #address-cells and #size-cells, the parent would need to
be overlaid, which is a bad tradeoff as it's either redundant or it
overrides the existing cell count with the wrong ones, possibly
invalidating reg parsing for existing nodes.

Improve the situation and avoid false positives by silencing the warnings
when building device tree overlays.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231130070320.3295372-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>