ARM: i.MX8M: delete unused per-board ddr_ddrphy_trained_csr array
ddrphy_trained_csr is an array of register/value pairs. The registers
are defined at compile time and the values are read back from the DDR
controller at runtime to be saved into another dram_timing_info in
dram_config_save. The

The arrays are generated for every board by the DDR tool, but neither
barebox or the original U-Boot code make use of this. Instead a global
ddrphy_trained_csr array is used for that purpose and the arrays in the
board code, just eat up space.

This waste of space is especially problematic for the i.MX8MN-EVK image,
which has both DDR4 and LPDDR4 firmware and their timings and overflows
the OCRAM now. Let's drop the unused arrays for a saving of 7K in the
i.MX8MN-EVK image.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231205083604.1497018-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>