of: define macro for maximum of_device_is_compatible return value
of_device_is_compatible() will return zero on mismatch and a positive
value up to and including (INT_MAX / 2) on match. Knowledge of the upper
bound can be useful to callers, because if inside a loop a maximum score
was achieved, there's usually no point in continuing the search.

Therefore add a macro for INT_MAX / 2 and document it for potential
users.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231205113618.3966168-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>