bootchooser: add new value "reset" to reset boot attempts counter
bootchooser provides a mechanism that allows for resetting the
"remaining_attempts" counters of all enabled boot slots to their
defaults.

The current possible trigger values to be stored in the defining
variable $bootchooser.reset_attempts are "all-zero" (i.e. no attempts
left at any boot slot) and "power-on" (i.e. $global.system.reset="POR").

We now want to add the option "reset" (i.e. $global.system.reset="RST")
to that list, so that we do not have to perform an entire power cycle
every time we need a counter reset.

By introducing the "reset" value as an option, we can deliberately cause
the bootchooser to set the "remaining_attempts" values to their defaults
when performing a generic restart. The restart handler of the main OS
is hence responsible that the subsequent barebox run sees "RESET_RST" as
reset reason instead of e.g. "RESET_WDG".

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Link: https://lore.barebox.org/20231214190125.1362776-1-h.assmann@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>