crypto: caam - remove overly verbose debug print
We can already read /dev/hwrng to get the randomness as generated by the
device. Printing 32K bytes of randomness to the log (or even worse to
console) is just not useful for debugging. There are debug prints for
the job descriptors and Linux doesn't have this print too since commit
2c5e88dc90f5 ("crypto: caam - simplify RNG implementation"), so let's
drop it in barebox too.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240102165537.1378708-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>