nvmem: ocotp: correct regmap's max_register
The Fuse words of the i.MX are 32-bit wide each and their number vary by
SoC. For the i.MX, there are 128 fuse words, which if linearized, are a
total 512 bytes in size.

The barebox driver calls the total size num_regs, which is misleading
and then sets the max_register to be the total_size - 1, which is wrong
as it's supposed to be e a multiple of the register stride, which is not
the case for (512 / 4) - 1 == 127.

Instead, we should be setting 508. Fix the calculation to do this and
while at it, rename the driver data member to its Linux counterpart.

Fixes: 2ee6bb35043d ("ARM: i.MX: ocotp: Switch to regmap support")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240102170100.1596372-6-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>