ARM: Layerscape: LS1028a: add psci node
The upstream dts files do not contain a psci node. Add one in order
to use PSCI to reset the system. This will also add a PSCI node to the
dtb the kernel is started with so that the Kernel can use PSCI to bring
up the secondary CPU core.

Link: https://lore.barebox.org/20240109161527.3237581-14-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>