mtd: nand: denali: Make partition binding configurable
The binding for the denali NAND driver changed over time. Originally the
partitions were placed directly under the controller node. Nowadays
there is an additional chip node between the controller node and the
partitions. So what originally was:

nand: nand@ff900000 {
	compatible = "altr,socfpga-denali-nand";
	partitions {
		...
	};
};

has now become:

nand: nand@ff900000 {
	compatible = "altr,socfpga-denali-nand";
	nand@0 {
		partitions {
			...
		};
	};
};

The new binding has been introduced with Linux-5.2 and after a grace
period it has been made mandatory in Linux-5.5. We want however to be
able to start old Kernels supporting only the old binding as well, so
make the binding we fix up the kernel device tree with configurable.

This introduces the globalvar nandx.denali_partition_binding with three
different settings:

chip: Fixup the Kernel device tree according to the new binding. This is
      the default
controller: Fixup the Kernel device tree according to the old binding
auto: When the Kernel device tree has a "nand@0" node then use the new
      binding, otherwise use the old binding.

The previous (only) implemented default before this patch was "auto".
This deliberately is changed with this patch as Kernels older than
Linux-5.2 are quite old already and likely contains the least surprises
for the user.

Link: https://lore.barebox.org/20240110131557.1289602-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>