ARM: dts: layerscape: mark ls1046a SoC DMA incoherent in DT
This effectively reverts in the barebox DT, Linux commit 136975c33894
("arm64: dts: ls1046a: make dma-coherent global to the SoC").

The reason for that is while the CCI-400 can be configured to make all
bus masters DMA-coherent, barebox (and older versions of U-Boot) didn't yet
do so. We will rectify this in a later commit, but for now restore
working order in case Linux is booted with the barebox DT.

Fixes: 32e2176ba050 ("dts: update to v6.1-rc1")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240110160112.4134162-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>